+++ frontend/src/App.jsx
@@ -8,6 +8,7 @@ import Account from './pages/Account.jsx'
 import Rapporten from './pages/Rapporten.jsx'
 import Meldingen from './pages/Meldingen.jsx'
 import Garantie from './pages/Garantie.jsx'
+import Facturen from './pages/Facturen.jsx'
 import Admin from './pages/Admin.jsx'
 import Webshop from './pages/Webshop.jsx'
 import Apps from './pages/Apps.jsx'
@@ -400,6 +401,7 @@ export default function App() {
         <Route path="/meldingen" element={<Meldingen />} />
         <Route path="/support" element={<Support user={user} />} />
         <Route path="/garantie" element={<Garantie />} />
+        <Route path="/facturen" element={<Facturen />} />
         <Route path="/account" element={<Account user={user} onLogout={onLogout} />} />
         <Route path="/admin"        element={<Admin />} />
         <Route path="/onderhoud"    element={<Onderhoud />} />